requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
click>=8.1.0
colorama>=0.4.6
sympy>=1.12
numpy>=1.24.0
matplotlib>=3.7.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0 